Score in fp32: 70% of every top-100 was sitting in an exact-score tie
Phase 5's sweep would not reproduce the 1a baseline at high parent weight, and
chasing that led to a real bug rather than a rounding quirk.

The 4c search stored embeddings in fp16 (fine, a storage decision) but also
produced the SCORES in fp16. Near 0.78 the representable spacing is 2^-11 =
0.000488, which is exactly the granularity visible in the run files. Measured
across all 13 domains: 69.9% of retrieved documents sit in an exact-score tie
group, with only ~50 distinct scores per top-100 (bitcoin 82.5%). For most of
the list the ranking was decided by the corpus-order tie-break instead of by
the model. Casting the matmul to fp32 costs 820 MB for History's 200k unique
vectors against 15.6 GB available. Re-running.

Two corrections to the sweep itself, both mine.

The high-w column is not the baseline. Because ~70% of parent scores are tied,
an infinitesimal step contribution still reorders tied documents, so w=1e6
gives 0.2572 against the 1a run's measured 0.2567. The script now takes
--baseline and compares against the separately scored 1a run, printing the
anchor beside it so the gap stays visible rather than being silently absorbed.

And w=0 is NOT TEMPO's Step-Only condition. I had labelled it that and told the
reader to expect it to lose badly against 14.6. TEMPO's Step-Only retrieves
with the step text alone; our 1b runs use the official Query+Step template, so
every step list already contains the full query. w=0 removes redundancy, not
context, which is why it scored 0.2592 rather than collapsing. The docstring
and the closing note now say so.

First sweep results, pending the fp32 re-run: the curve peaks at w=0.25 with
0.2671 against the 1a run's 0.2567, +0.0099 with a paired CI of [+0.0060,
+0.0140], and stays significant through w=5. Small but real, and the shape is
monotone from the peak to the anchor rather than noisy.

+++ kaggle/kernels/phase4c_dense_search/phase4c_dense_search.py
@@ -185,7 +185,14 @@ def main() -> int:
         if vectors.shape[0] != int(meta["n_unique"]):
             die(f"{domain}: {vectors.shape[0]} vectors but doc_index says {meta['n_unique']}")
 
-        mat = torch.from_numpy(vectors).cuda().half()          # [n_unique, dim]
+        # fp32 for the similarity matmul, not fp16. Storing the embeddings in fp16 is fine
+        # -- that is a storage decision -- but producing the SCORES in fp16 quantises them
+        # to a 2^-11 grid near 0.78, and run 3 measured **69.9% of retrieved documents
+        # sitting in an exact-score tie group**, only ~50 distinct scores per top-100. The
+        # ranking was being decided by the corpus-order tie-break rather than by the model
+        # for most of the list. fp32 costs 820 MB for History's 200k unique vectors against
+        # 15.6 GB available, which is nothing.
+        mat = torch.from_numpy(vectors).cuda().float()         # [n_unique, dim]
         rows_t = torch.from_numpy(rows).cuda()
 
         # ---- 1a: whole query -------------------------------------------------
@@ -206,13 +213,13 @@ def main() -> int:
             if not items:
                 continue
             ids = [i for i, _ in items]
-            qv = torch.from_numpy(encode_queries([t for _, t in items])).cuda().half()
+            qv = torch.from_numpy(encode_queries([t for _, t in items])).cuda().float()
             d_runs = RUNS / domain
             d_runs.mkdir(parents=True, exist_ok=True)
             path = d_runs / f"run_{sub}_{SPLIT}.trec"
             with path.open("w", encoding="utf-8", newline="\n") as fh:
                 for start in range(0, len(ids), 64):
-                    sims = (qv[start:start + 64] @ mat.T).float()      # [b, n_unique]
+                    sims = qv[start:start + 64] @ mat.T               # [b, n_unique], fp32
                     per_doc = sims[:, rows_t].cpu().numpy()            # [b, n_docs]
                     for i, qid in enumerate(ids[start:start + 64]):
                         scores = per_doc[i]
